requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
jalali_core==1.0.0
jdatetime==5.0.0
Markdown==3.5.2
psycopg==3.1.19
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.12.0
